apps/backend/requirements.txt
aiohttp==3.11.11
fastapi==0.115.8
feedparser==6.0.11
numpy==2.2.2
passlib==1.7.4
pgvector==0.3.6
pydantic==2.10.6
python_dateutil==2.9.0.post0
python_jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.15.1
sentence_transformers==3.4.0
SQLAlchemy==2.0.37
torch==2.5.1
transformers==4.48.1
uvicorn==0.34.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
lxml==5.3.0
lxml_html_clean==0.4.1
bcrypt==3.2.2
goose3==3.1.19
tqdm
python-dotenv
langchain
langchain-core
langchain-google-genai
google-generativeai